+++ app/sprinkles/account/assets/userfrosting/js/pages/forgot-password.js
@@ -2,7 +2,7 @@
  * Page-specific Javascript file.  Should generally be included as a separate asset bundle in your page template.
  * example: {{ assets.js('js/pages/sign-in-or-register') | raw }}
  *
- * This script depends on validation rules specified in components/page.js.twig.
+ * This script depends on validation rules specified in pages/partials/page.js.twig.
  *
  * Target page: account/forgot-password
  */